Move all arg structs above their function
