+++ api/exits.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/hermeznetwork/hermez-node/api/requests"
 	"github.com/hermeznetwork/hermez-node/db/historydb"
 )
 
@@ -36,7 +35,7 @@ func (a *API) getExits(c *gin.Context) {
 	}
 
 	// Fetch exits from historyDB
-	exits, pendingItems, err := a.h.GetExitsAPI(requests.GetExitsAPIRequest{
+	exits, pendingItems, err := a.h.GetExitsAPI(historydb.GetExitsAPIRequest{
 		EthAddr:              addr,
 		Bjj:                  bjj,
 		TokenID:              tokenID,
